+++ app/src/main/java/dev/blunch/blunch/activity/GetPaymentMenuActivity.java
@@ -110,7 +110,7 @@ public class GetPaymentMenuActivity extends AppCompatActivity {
         TextView messageCount = (TextView) findViewById(R.id.messagesCount);
 
         if(guest() || host()) {
-            int count = menuService.getPendingMessagesCount(this.menuId);
+            long count = menuService.getPendingMessagesCount(this.menuId);
             if (count > 0) {
                 if (count < 100) messageCount.setText(String.valueOf(count));
                 else messageCount.setText("+99");
